feat(template-x): api and hover wip
